+++ reactivestate/src/commonMain/kotlin/com/ensody/reactivestate/AutoRunStateFlow.kt
@@ -34,5 +34,6 @@ private class StateFlowObservable<T>(
 
     override fun removeObserver() {
         observer?.cancel()
+        observer = null
     }
 }