+++ reactivestate/src/jvmTest/kotlin/com/ensody/reactivestate/ObserverTest.kt
@@ -50,15 +50,17 @@ internal class ObserverTest : CoroutineTest() {
 
     @Test
     fun derivedObservableOnCoroutineScope() = runBlockingTest {
-        val source = MutableValueFlow(0)
+        val rawSource = MutableValueFlow(0)
+        val source = scopelessDerived { get(rawSource) }
         lateinit var target: StateFlow<Int>
         lateinit var scopelessTarget: StateFlow<Int>
         val job = launch {
             target = derived { 2 * get(source) }
             scopelessTarget = scopelessDerived { 2 * get(source) }
             val lazyTarget = derived(0, Lazily) { 2 * get(source) }
-            val superLazyTarget = derived(0, WhileSubscribed()) { 2 * get(source) }
-            val superLazyScopelessTarget = derivedWhileSubscribed(0) { 2 * get(source) }
+            val superLazyTargetInner = derivedWhileSubscribed(0) { get(source) }
+            val superLazyTarget = derived(0, WhileSubscribed()) { 2 * get(superLazyTargetInner) }
+            val superLazyScopelessTarget = derivedWhileSubscribed(0) { 2 * get(superLazyTargetInner) }
             val asyncTarget = derived(-1, Eagerly) { delay(100); 2 * get(source) }
 
             // Right after creation of the derived observable the values should be in sync
@@ -69,15 +71,31 @@ internal class ObserverTest : CoroutineTest() {
             assertEquals(0, superLazyScopelessTarget.value)
             assertEquals(-1, asyncTarget.value)
 
+            // Once somebody collects the lazy derived flow, the value gets updated continuously
+            var superLazyJob = launch { superLazyTarget.collect() }
+            var superLazyScopelessJob = launch { superLazyScopelessTarget.collect() }
+            listOf(2, 5, 0).forEach {
+                rawSource.value = it
+                assertEquals(2 * it, target.value)
+                assertEquals(2 * it, scopelessTarget.value)
+                assertEquals(it, superLazyTargetInner.value)
+                assertEquals(2 * it, superLazyTarget.value)
+                assertEquals(2 * it, superLazyScopelessTarget.value)
+            }
+            // Now we stop collecting, so the values won't update anymore
+            superLazyJob.cancel()
+            superLazyScopelessJob.cancel()
+
             advanceTimeBy(120)
             assertEquals(0, asyncTarget.value)
 
             // Setting value multiple times should work
             listOf(2, 5, 10).forEach {
-                source.value = it
+                rawSource.value = it
                 assertEquals(2 * it, target.value)
                 assertEquals(2 * it, scopelessTarget.value)
                 assertEquals(0, lazyTarget.value)
+                assertEquals(0, superLazyTargetInner.value)
                 assertEquals(0, superLazyTarget.value)
                 assertEquals(0, superLazyScopelessTarget.value)
             }
@@ -88,20 +106,23 @@ internal class ObserverTest : CoroutineTest() {
             assertEquals(2 * 10, asyncTarget.value)
 
             val lazyJob = launch { lazyTarget.collect() }
-            val superLazyJob = launch { superLazyTarget.collect() }
-            val superLazyScopelessJob = launch { superLazyScopelessTarget.collect() }
+            // Now we re-subscribe, so the value should get updates again
+            superLazyJob = launch { superLazyTarget.collect() }
+            superLazyScopelessJob = launch { superLazyScopelessTarget.collect() }
             advanceUntilIdle()
+            lazyJob.cancel()
 
             // Once somebody collects the lazy derived flow, the value gets updated continuously
             assertEquals(2 * 10, lazyTarget.value)
+
             listOf(2, 5, 10).forEach {
-                source.value = it
+                rawSource.value = it
                 assertEquals(2 * it, lazyTarget.value)
+                assertEquals(it, superLazyTargetInner.value)
                 assertEquals(2 * it, superLazyTarget.value)
                 assertEquals(2 * it, superLazyScopelessTarget.value)
             }
 
-            lazyJob.cancel()
             superLazyJob.cancel()
             superLazyScopelessJob.cancel()
 
@@ -109,10 +130,11 @@ internal class ObserverTest : CoroutineTest() {
             assertEquals(2 * 10, superLazyTarget.value)
             assertEquals(2 * 10, superLazyScopelessTarget.value)
             listOf(2, 5, 10).forEach {
-                source.value = it
+                rawSource.value = it
                 // Even when nobody is listening anymore, we continue updating
                 assertEquals(2 * it, lazyTarget.value)
                 // We stop updating when nobody listens
+                assertEquals(10, superLazyTargetInner.value)
                 assertEquals(2 * 10, superLazyTarget.value)
                 assertEquals(2 * 10, superLazyScopelessTarget.value)
             }
@@ -121,7 +143,7 @@ internal class ObserverTest : CoroutineTest() {
         }
         job.join()
         val oldValue = source.value * 2
-        source.value += 5
+        rawSource.value += 5
         assertEquals(oldValue, target.value)
         assertEquals(oldValue + 10, scopelessTarget.value)
     }